Add Event dispatcher provider
