+++ src/Bundle/Resources/config/services/state.xml
@@ -24,9 +24,17 @@
 
         <service id="sylius.state.provider" class="Sylius\Component\Resource\State\Provider">
             <argument type="tagged_locator" tag="sylius.state_provider" />
-            <argument type="service" id="sylius.dispatcher.operation" />
         </service>
         <service id="Sylius\Component\Resource\State\ProviderInterface" alias="sylius.state.provider" />
+
+        <service id="sylius.state.provider.event_dispatcher"
+                 class="Sylius\Component\Resource\State\EventDispatcherProvider"
+                 decorates="sylius.state.provider"
+                 decoration-priority="-32"
+        >
+            <argument type="service" id=".inner" />
+            <argument type="service" id="sylius.dispatcher.operation" />
+        </service>
 
         <service id="sylius.state.factory" class="Sylius\Component\Resource\State\Factory">
             <argument type="tagged_locator" tag="sylius.resource_factory" />
