+++ src/Component/State/Provider.php
@@ -15,10 +15,7 @@ namespace Sylius\Component\Resource\State;
 
 use Psr\Container\ContainerInterface;
 use Sylius\Component\Resource\Context\Context;
-use Sylius\Component\Resource\Metadata\CollectionOperationInterface;
 use Sylius\Component\Resource\Metadata\Operation;
-use Sylius\Component\Resource\Metadata\ShowOperationInterface;
-use Sylius\Component\Resource\Symfony\EventDispatcher\OperationEventDispatcherInterface;
 use Webmozart\Assert\Assert;
 
 /**
@@ -28,7 +25,6 @@ final class Provider implements ProviderInterface
 {
     public function __construct(
         private ContainerInterface $locator,
-        private OperationEventDispatcherInterface $operationEventDispatcher,
     ) {
     }
 
@@ -38,13 +34,6 @@ final class Provider implements ProviderInterface
 
         if (null === $provider) {
             return null;
-        }
-
-        if (
-            $operation instanceof CollectionOperationInterface ||
-            $operation instanceof ShowOperationInterface
-        ) {
-            $this->operationEventDispatcher->dispatch(null, $operation, $context);
         }
 
         if (\is_callable($provider)) {
